SEN-56 : Automate consumption generation in function of the projects in the csv file (#52)
* Update data deletion handling

The ck and joular data are deleted accordingly to the project. Until now, all the data were deleted each time I wanted to add new data to the db, but from now on, only the remaining data of the project I want to add will be deleted. The other data wont be touched

* Adding logs